Split visual component out of MusicController
